+++ src/RunCli/Command/ModelCommand.php
@@ -1,6 +1,7 @@
 <?php
 namespace RunCli\Command;
 
+use RunCli\CliTrait;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -9,6 +10,8 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class ModelCommand extends Command
 {
+    use CliTrait;
+
     private $modelName;
     private $methods;